pyproject.toml
[project]
name = "windi messenger"
version = "0.1.0"
description = ""
authors = [
    {name = "nickpashkevich",email = "micolas9441@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "ruff (>=0.11.9,<0.12.0)",
    "pyhumps (>=3.8.0,<4.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "black (>=25.1.0,<26.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "alembic-postgresql-enum (>=1.7.0,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*"
]